fix: Исправления после код-ревью
